# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/reports/layouts": {
      "get": {
        "operationId": "ReportingController_listLayouts",
        "parameters": [
          {
            "name": "report_key",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reporting"
        ]
      },
      "post": {
        "operationId": "ReportingController_createLayout",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Reporting"
        ]
      }
    },
    "/api/v1/reports/layouts/{id}": {
      "patch": {
        "operationId": "ReportingController_updateLayout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reporting"
        ]
      },
      "delete": {
        "operationId": "ReportingController_deleteLayout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reporting"
        ]
      }
    },
    "/api/v1/reports/print/invoices/{id}": {
      "get": {
        "operationId": "ReportingController_invoicePrint",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reporting"
        ]
      }
    },
    "/api/v1/reports/print/purchase-invoices/{id}": {
      "get": {
        "operationId": "ReportingController_purchaseInvoicePrint",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reporting"
        ]
      }
    },
    "/api/v1/reports/print/vouchers/{id}": {
      "get": {
        "operationId": "ReportingController_voucherPrint",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reporting"
        ]
      }
    },
    "/api/v1/reports/print/journal-entries/{id}": {
      "get": {
        "operationId": "ReportingController_journalPrint",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reporting"
        ]
      }
    },
    "/api/v1/reports/print/shifts/{id}": {
      "get": {
        "operationId": "ReportingController_shiftPrint",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reporting"
        ]
      }
    },
    "/api/v1/reports/print/inventory-documents/{id}": {
      "get": {
        "operationId": "ReportingController_inventoryDocumentPrint",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reporting"
        ]
      }
    },
    "/api/v1/reports/{key}": {
      "get": {
        "operationId": "ReportingController_run",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reporting"
        ]
      }
    },
    "/api/v1/reports/{key}/export": {
      "post": {
        "operationId": "ReportingController_export",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Reporting"
        ]
      }
    },
...
    "/api/v1/pos/held-tickets": {
      "get": {
        "operationId": "PosController_heldTickets",
        "parameters": [
          {
            "name": "branch_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "history",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      },
      "post": {
        "operationId": "PosController_holdTicket",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/api/v1/pos/held-tickets/{id}/recall": {
      "post": {
        "operationId": "PosController_recallHeldTicket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/api/v1/pos/held-tickets/{id}/cancel": {
      "post": {
        "operationId": "PosController_cancelHeldTicket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/api/v1/pos/shortcuts": {
      "get": {
        "operationId": "PosController_shortcuts",
        "parameters": [
          {
            "name": "branch_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      },
      "post": {
        "operationId": "PosController_setShortcut",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/api/v1/pos/shortcuts/{slot}/remove": {
      "post": {
        "operationId": "PosController_removeShortcut",
        "parameters": [
          {
            "name": "slot",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/api/v1/pos/categories": {
      "get": {
        "operationId": "PosController_categories",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      },
      "post": {
        "operationId": "PosController_createCategory",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/api/v1/pos/tables": {
      "get": {
        "operationId": "PosController_tables",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      },
      "post": {
        "operationId": "PosController_createTable",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/api/v1/pos/tables/{id}/open": {
      "post": {
        "operationId": "PosController_open",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/api/v1/pos/tables/{id}/items": {
      "post": {
        "operationId": "PosController_addItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/api/v1/pos/events/{id}/void": {
      "post": {
        "operationId": "PosController_voidItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/api/v1/pos/tables/{id}/send-to-invoice": {
      "post": {
        "operationId": "PosController_send",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/api/v1/pos/tables/{id}/close": {
      "post": {
        "operationId": "PosController_close",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
...
    "/api/v1/pos/tables/{sourceId}/merge/{targetId}": {
      "post": {
        "operationId": "PosController_merge",
        "parameters": [
          {
            "name": "sourceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/api/v1/pos/tables/{id}/split": {
      "post": {
        "operationId": "PosController_split",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
...
  },
  "info": {
    "title": "ERP SaaS API",
    "description": "Cloud multi-tenant ERP. Errors are RFC 9457 application/problem+json with a stable `code` from @erp/contracts. Tenant context always comes from the access token.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "branchScope": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Branch-Id"
      },
      "idempotencyKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Idempotency-Key"
      }
    },
...
  }
}